Develop

+++ projects/local-private-network/compile-geth.sh
@@ -1,17 +1,85 @@
 #!/bin/bash
 set -e
 
+showHelp() {
+# `cat << EOF` This means that cat should stop reading when EOF is detected
+cat << EOF
+Usage: ./wrapper.sh -e <compiler-env>
+Spin up Foundry with Geth and a database.
+
+-h,         Display help
+
+-e,         Should we compile on your "local" machine or in "docker" or on a "remote" server
+
+-v,         Should we "remove" the volume when bringind the image down or "keep" it?
+
+-u,         What username should we use for the remote build?
+
+EOF
+exit 1
+# EOF is found above and hence cat command stops reading. This is equivalent to echo but much neater when printing out.
+}
+
+u="abdul"
+n="alabaster.lan.vdb.to"
+while getopts ":e:u:n:" o; do
+    case "${o}" in
+        e)
+            e=${OPTARG}
+            [ "$e" = "local" -o "$e" = "docker" -o "$e" = "remote" ] || showHelp
+            ;;
+        u)
+            u=${OPTARG}
+            ;;
+        n)
+            n=${OPTARG}
+            ;;
+        *)
+            showHelp
+            ;;
+    esac
+done
+shift $((OPTIND-1))
+
 GREEN='\033[0;32m'
 RED='\033[0;31m'
 NC='\033[0m'
 
 start_path=$(pwd)
+
+trap "cd ${start_path}" SIGINT SIGTERM
+
 cd ../../../../
 echo -e "${GREEN}Building geth!${NC}"
-docker build -t vulcanize/go-ethereum -f Dockerfile .
-docker run --rm --entrypoint cat vulcanize/go-ethereum /usr/local/bin/geth > ./geth-linux-amd64
-chmod +x ./geth-linux-amd64
+echo -e "${GREEN}e=${e}${NC}"
+echo -e "${GREEN}n=${n}${NC}"
+echo -e "${GREEN}u=${u}${NC}"
 
-echo -e "${GREEN}geth build complete!${NC}"
+if [[ "$e" = "local" ]] ; then
+    echo -e "${GREEN}Building geth locally!${NC}"
+    echo -e "${RED}LOCAL BUILD MIGHT NOT WORK!!${NC}"
+    make geth
+    chmod +x build/bin/geth
+    mv build/bin/geth related-repositories/foundry-test/projects/local-private-network/geth-linux-amd64
+fi
+
+if [[ "$e" = "docker" ]] ; then
+    echo -e "${GREEN}Building geth using docker!${NC}"
+    docker build -t vulcanize/go-ethereum -f Dockerfile .
+    docker run --rm --entrypoint cat vulcanize/go-ethereum /usr/local/bin/geth > ./geth-linux-amd64
+    chmod +x ./geth-linux-amd64
+    mv ./geth-linux-amd64 related-repositories/foundry-test/projects/local-private-network/geth-linux-amd64
+fi
+
+if [[ "$e" == "remote" ]]; then
+    echo -e "${GREEN}Building geth remotely on $n ${NC}"
+    [ -e related-repositories/foundry-test/projects/local-private-network/geth-linux-amd64 ] && \
+        rm related-repositories/foundry-test/projects/local-private-network/geth-linux-amd64
+    rsync -uavz ./ ${u}@${n}:/home/${u}/go-ethereum-cerc
+    ssh ${u}@${n} "cd /home/${u}/go-ethereum-cerc/ ; make geth ; chmod +x build/bin/geth"
+    scp ${u}@${n}:/home/${u}/go-ethereum-cerc/build/bin/geth related-repositories/foundry-test/projects/local-private-network/geth-linux-amd64
+
+fi
+
 cd $start_path
-mv ../../../../geth-linux-amd64 .
+echo -e "${GREEN}geth build complete!${NC}"
